# public/docs/messenger-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Jugajug Enterprise Real-Time Messenger, Calling & Communication API",
    "version": "1.0.0",
    "description": "Comprehensive, production-ready, API-first communication platform specification for Jugajug. Supports 1-to-1, group messaging, WebRTC audio/video calling, voice notes, pinning, saving, multi-device sessions, and offline event replay synchronization for Web, Android, iOS, and Desktop apps."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000/api/v1",
      "description": "Local Development Server"
    },
    {
      "url": "https://jugajug.com/api/v1",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum Token"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/devices/register": {
      "post": {
        "summary": "Register or update current device session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["device_id"],
                "properties": {
                  "device_id": { "type": "string" },
                  "device_name": { "type": "string" },
                  "platform": { "type": "string", "enum": ["web", "android", "ios", "desktop", "windows", "macos", "linux"] },
                  "app_version": { "type": "string" },
                  "push_token": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Device registered." }
        }
      }
    },
    "/devices/{id}": {
      "delete": {
        "summary": "Revoke a device session",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Session revoked." }
        }
      }
    },
...
  }
}